+++ framework/Furion.Pure/RemoteRequest/Extensions/RemoteRequestStringExtensions.cs
@@ -187,7 +187,7 @@ namespace Furion.RemoteRequest.Extensions
         }
 
         /// <summary>
-        /// 设置超时时间（分钟）
+        /// 设置超时时间（秒）
         /// </summary>
         /// <param name="requestUrl"></param>
         /// <param name="timeout"></param>