+++ framework/Furion.Pure/RemoteRequest/Internal/HttpClientExecutePart.cs
@@ -74,7 +74,7 @@ namespace Furion.RemoteRequest
         public List<(string Name, byte[] Bytes, string FileName)> BodyBytes { get; private set; } = new List<(string Name, byte[] Bytes, string FileName)>();
 
         /// <summary>
-        /// 超时时间（分钟）
+        /// 超时时间（秒）
         /// </summary>
         public long Timeout { get; private set; } = 0;